backend/requirements.txt
fastapi
uvicorn
pymongo
qdrant-client
python-multipart
pydantic
numpy
Pillow
tqdm
sentence-transformers
torch
tinydb
